requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async and concurrent processing
aiohttp>=3.8.0

# Logging and monitoring
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: for enhanced charts and visualizations
plotly>=5.15.0
altair>=5.0.0